move button ref into button module
