+++ roller/src/control/control_mapping.rs
@@ -1,42 +1,14 @@
 use roller_protocol::FaderId;
 use rustc_hash::FxHashMap;
-use std::time::Instant;
 
 use roller_protocol::{ButtonCoordinate, ButtonGridLocation};
 
 use crate::{
     control::{
-        button::{ButtonGroup, ButtonMapping, MetaButtonMapping, PadMapping},
+        button::{ButtonGroup, ButtonMapping, MetaButtonMapping, PadMapping, ButtonRef},
         fader::FaderControlMapping,
-        NoteState,
     },
-    lighting_engine::ControlEvent,
 };
-
-pub enum ButtonRef<'a> {
-    Standard(&'a ButtonGroup, &'a ButtonMapping),
-    Meta(&'a MetaButtonMapping),
-}
-impl<'a> ButtonRef<'a> {
-    pub fn into_control_event(
-        self,
-        note_state: NoteState,
-        now: Instant,
-    ) -> Option<ControlEvent<'a>> {
-        match (self, note_state) {
-            (ButtonRef::Standard(group, button), _) => {
-                Some(ControlEvent::UpdateButton(group, button, note_state, now))
-            }
-            (ButtonRef::Meta(meta_button), NoteState::On) => {
-                Some(meta_button.on_action.control_event(now))
-            }
-            (ButtonRef::Meta(meta_button), NoteState::Off) => meta_button
-                .off_action
-                .as_ref()
-                .map(|action| action.control_event(now)),
-        }
-    }
-}
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct ControlMapping {
